refactor : 자동이체 로직 수정
