+++ src/main/java/com/heeha/HanaHeritageBeApplication.java
@@ -1,11 +1,14 @@
 package com.heeha;
 
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableScheduling
 public class HanaHeritageBeApplication {
 
     public static void main(String[] args) {
