--- src/main/java/com/heeha/domain/account/jihu/dto/AccountCheckResponse.java
@@ -1,25 +1,0 @@
-package com.heeha.domain.account.jihu.dto;
-
-import java.time.LocalDateTime;
-
-import com.heeha.domain.account.entity.Account;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class AccountCheckResponse {
-    private Long id;
-    private Long accountNumber;
-    private String name;
-    private Long balance;
-    private LocalDateTime createdAt;
-
-    public AccountCheckResponse(Account account) {
-        this.id = account.getId();
-        this.accountNumber = account.getAccountNumber();
-        this.name = account.getName();
-        this.balance = account.getBalance();
-        this.createdAt = account.getCreated_at();
-    }
-}
