--- src/main/java/com/heeha/domain/account/jihu/repository/AccountRepository.java
@@ -1,16 +1,0 @@
-package com.heeha.domain.account.jihu.repository;
-
-
-import com.heeha.domain.account.entity.Account;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
-    Account findByAccountNumber(Long accountNumber);
-    @Query("SELECT a FROM account a join fetch a.customer where a.customer.id = :customerId")
-    List<Account> findAccountFixByCustomerId(Long customerId);
-}
