+++ src/main/resources/application-dev.yml
@@ -13,12 +13,12 @@ spring:
     redis:
       host: localhost
       port: 6379
-      password: 111111
+      password: aaaa
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://127.0.0.1:3306/heritagedb?characterEncoding=utf8
-    username: heeha
-    password: '0612'
+    username: heritage
+    password: 123123
   jpa:
     hibernate:
       ddl-auto: update
@@ -34,6 +34,11 @@ logging:
 jwt:
   secret: 0mUet36qbTmKpG6lQPJJbij815G75KB2zwUim2U6Hjc=
   issuer: quswjdgma83@naver.com
-
 app:
   key: ebfd212c9fc8a2c8774d9dee8c034bd7
+batch:
+  jdbc:
+    initialize-schema: never
+    # initialize-schema: always
+  job:
+    enabled: false # 자동 실행 방지
